Split handler file
