+++ go/cert_srv/internal/drkey/BUILD.bazel
@@ -3,7 +3,9 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = [
-        "handler.go",
+        "lvl1repHandler.go",
+        "lvl1reqHandler.go",
+        "lvl2reqHandler.go",
         "requester.go",
         "store_keeper.go",
         "testcommon.go",
